isolated code into methods
