+++ web-faucet/src/lib/Form.svelte
@@ -2,53 +2,39 @@
   import Tick from "./icons/Tick.svelte";
   import Cross from "./icons/Cross.svelte";
   import {doRecaptcha} from "./captcha";
+  import {createEventDispatcher} from "svelte";
+  import {CAPTCHA_KEY} from "./utils/config";
+  import {faucetRequest} from "./utils/faucetRequest";
+  import {fly} from 'svelte/transition';
 
-  const captchaKey: string = import.meta.env.VITE_CAPTCHA_KEY;
-  const faucetUrl = import.meta.env.VITE_FAUCET_URL;
-
+  const dispatch = createEventDispatcher<{ submit: string }>();
 
   let address: string = '';
 
   let webRequest: Promise<string> = null;
 
   function onSubmit() {
-    webRequest = boilerplateRequest(address);
+    webRequest = request(address);
+    dispatch('submit', address);
   }
 
-  async function boilerplateRequest(address: string): Promise<string> {
+  async function request(address: string): Promise<string> {
+    const token = await doRecaptcha(CAPTCHA_KEY);
+    return boilerplateRequest(address, token);
+  }
+
+  async function boilerplateRequest(address: string, token: string): Promise<string> {
     await new Promise((resolve) => setTimeout(resolve, 1000));
-    const token = await doRecaptcha(captchaKey);
     if (address === "error") {
       throw new Error("This is a terrible error!");
     }
     console.log(token);
     return "0x7824400bf61a99c51b946454376a84c636a2d86070996a6a5f55999b26e7df51";
   }
-
-  async function faucetRequest(address: string): Promise<string> {
-    const token = await doRecaptcha(captchaKey);
-    const body = {
-      address,
-      parachain_id: "1002",
-      recaptcha: token
-    }
-    const fetchResult = await fetch(faucetUrl, {
-      method: "POST", body: JSON.stringify(body), headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-    });
-    const result = await fetchResult.json();
-    if ('error' in result) {
-      throw new Error(result.error);
-    } else {
-      return result.hash;
-    }
-  }
 </script>
 
 <svelte:head>
-  <script src="https://www.google.com/recaptcha/api.js?render={captchaKey}" async defer></script>
+  <script src="https://www.google.com/recaptcha/api.js?render={CAPTCHA_KEY}" async defer></script>
 </svelte:head>
 
 <form on:submit|preventDefault={onSubmit}>
