+++ web-faucet/src/lib/utils/config.ts
@@ -1,0 +1,2 @@
+export const CAPTCHA_KEY: string = import.meta.env.VITE_CAPTCHA_KEY;
+export const FAUCET_URL = import.meta.env.VITE_FAUCET_URL;
